# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A small in-memory CRUD API for managing a to-do list."
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "paths": {
    "/tasks": {
      "get": {
        "summary": "List all tasks (optionally filter/search)",
        "parameters": [
          { "name": "done", "in": "query", "required": false, "schema": { "type": "boolean" }, "description": "Filter by completion status" },
          { "name": "search", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Case-insensitive substring match on title" }
        ],
        "responses": {
          "200": {
            "description": "Array of tasks",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Task" } }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewTask" },
              "example": { "title": "Buy milk" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } }
          },
          "400": {
            "description": "Missing or empty title",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Get task statistics",
        "responses": {
          "200": {
            "description": "Counts of total/done/open tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": { "type": "integer" },
                    "done": { "type": "integer" },
                    "open": { "type": "integer" }
                  }
                },
                "example": { "total": 7, "done": 3, "open": 4 }
              }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "summary": "Reset tasks back to the 3 seed examples",
        "responses": {
          "200": {
            "description": "Tasks reset",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a single task by id",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "The task",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } }
          },
          "404": {
            "description": "Task not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      },
      "put": {
        "summary": "Update a task's title and/or done status",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTask" },
              "example": { "done": true }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated task",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } }
          },
          "400": {
            "description": "Invalid body",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": {
            "description": "Task not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "204": { "description": "Deleted, no content" },
          "404": {
            "description": "Task not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "title": { "type": "string", "example": "Buy milk" },
          "done": { "type": "boolean", "example": false }
        }
      },
      "NewTask": {
        "type": "object",
        "required": ["title"],
        "properties": { "title": { "type": "string", "example": "Buy milk" } }
      },
      "UpdateTask": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "example": "Buy oat milk" },
          "done": { "type": "boolean", "example": true }
        }
      },
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string", "example": "Task 99 not found" } }
      }
    }
  }
}