# server.js
const express = require('express');
const swaggerUi = require('swagger-ui-express');
const openapiSpec = require('./openapi.json');
const app = express();
const PORT = 3000;

app.use(express.json()); // parses JSON request bodies into req.body
app.use('/docs', swaggerUi.serve, swaggerUi.setup(openapiSpec));

// In-memory "database" — resets every time the server restarts.
let tasks = [
  { id: 1, title: 'Buy milk', done: false },
  { id: 2, title: 'Write README', done: false },
  { id: 3, title: 'Walk the dog', done: true }
];
let nextId = 4;

app.get('/', (req, res) => {
  res.json({
    name: 'Task API',
    version: '1.0',
    endpoints: ['/tasks']
  });
});

app.get('/health', (req, res) => {
  res.json({ status: 'ok' });
});

app.get('/tasks', (req, res) => {
  let result = tasks;

  // ?done=true or ?done=false — filter by completion status
  if (req.query.done !== undefined) {
    const wantDone = req.query.done === 'true';
    result = result.filter(t => t.done === wantDone);
  }

  // ?search=milk — case-insensitive substring match on title
  if (req.query.search) {
    const term = req.query.search.toLowerCase();
    result = result.filter(t => t.title.toLowerCase().includes(term));
  }

  res.json(result);
});

app.get('/stats', (req, res) => {
  const total = tasks.length;
  const done = tasks.filter(t => t.done).length;
  res.json({ total, done, open: total - done });
});

app.post('/reset', (req, res) => {
  tasks = [
    { id: 1, title: 'Buy milk', done: false },
    { id: 2, title: 'Write README', done: false },
    { id: 3, title: 'Walk the dog', done: true }
  ];
  nextId = 4;
  res.json({ message: 'Tasks reset to seed data', tasks });
});

app.get('/tasks/:id', (req, res) => {
  const id = Number(req.params.id);
  const task = tasks.find(t => t.id === id);
  if (!task) {
    return res.status(404).json({ error: `Task ${id} not found` });
  }
  res.json(task);
});

app.post('/tasks', (req, res) => {
  const { title } = req.body || {};

  if (!title || typeof title !== 'string' || title.trim() === '') {
    return res.status(400).json({ error: 'title is required and must be a non-empty string' });
  }

  const newTask = { id: nextId++, title: title.trim(), done: false };
  tasks.push(newTask);
  res.status(201).json(newTask);
});

app.put('/tasks/:id', (req, res) => {
  const id = Number(req.params.id);
  const task = tasks.find(t => t.id === id);
  if (!task) {
    return res.status(404).json({ error: `Task ${id} not found` });
  }

  const { title, done } = req.body || {};

  // At least one valid field must be provided, and if title is given it must be non-empty.
  const titleProvided = title !== undefined;
  const doneProvided = done !== undefined;

  if (!titleProvided && !doneProvided) {
    return res.status(400).json({ error: 'request body must include title and/or done' });
  }
  if (titleProvided && (typeof title !== 'string' || title.trim() === '')) {
    return res.status(400).json({ error: 'title must be a non-empty string' });
  }
  if (doneProvided && typeof done !== 'boolean') {
    return res.status(400).json({ error: 'done must be a boolean' });
  }

  if (titleProvided) task.title = title.trim();
  if (doneProvided) task.done = done;

  res.json(task);
});

app.delete('/tasks/:id', (req, res) => {
  const id = Number(req.params.id);
  const index = tasks.findIndex(t => t.id === id);
  if (index === -1) {
    return res.status(404).json({ error: `Task ${id} not found` });
  }
  tasks.splice(index, 1);
  res.status(204).send();
});

app.listen(PORT, () => {
  console.log(`Server running on http://localhost:${PORT}`);
});